Reorganize model and circuit classes
